// src/features/settings/domain/enums/settings-control-id.ts
export enum SettingsControlId {
  WIFI_ENABLED = 'wifi-enabled',
  WIFI_ASK = 'wifi-ask',
  BLUETOOTH_ENABLED = 'bluetooth-enabled',
  BLUETOOTH_AIRDROP = 'bluetooth-airdrop',
  NETWORK_FIREWALL = 'network-firewall',
  NETWORK_PRIVATE = 'network-private',
  NOTIFICATIONS_ALLOW = 'notifications-allow',
  NOTIFICATIONS_PREVIEW = 'notifications-preview',
  SOUND_INTERFACE = 'sound-interface',
  SOUND_FEEDBACK = 'sound-feedback',
  BATTERY_LOW_POWER = 'battery-low-power',
  BATTERY_OPTIMIZED = 'battery-optimized',
  FOCUS_DO_NOT_DISTURB = 'focus-dnd',
  FOCUS_SHARE = 'focus-share',
  GENERAL_UPDATES = 'general-updates',
  GENERAL_HANDOFF = 'general-handoff',
  DESKTOP_AUTO_HIDE = 'desktop-hide',
  DESKTOP_RECENTS = 'desktop-recents',
}